refactor(folders) moving charts components into charts/
