+++ frontend-web/src/components/charts/CategoryTotals.jsx
