+++ frontend-web/src/components/charts/LineChart.jsx
